+++ book/listings/css/3/style.css
@@ -1,3 +1,3 @@
-button.button_1 {
+button.button-1 {
   color: magenta;
 }
