+++ book/listings/css/4/main.rs
@@ -1,5 +1,7 @@
+use gdk::Display;
+use gtk::gdk;
 use gtk::prelude::*;
-use gtk::{Application, ApplicationWindow, Button};
+use gtk::{Application, ApplicationWindow, Button, CssProvider, StyleContext};
 
 fn main() {
     // Create a new application
@@ -8,20 +10,34 @@ fn main() {
         .build();
 
     // Connect to signals
+    app.connect_startup(|_| load_css());
     app.connect_activate(build_ui);
 
     // Run the application
     app.run();
 }
 
+fn load_css() {
+    // Load the CSS file and add it to the provider
+    let provider = CssProvider::new();
+    provider.load_from_data(include_bytes!("style.css"));
+
+    // Add the provider to the default screen
+    StyleContext::add_provider_for_display(
+        &Display::default().expect("Could not connect to a display."),
+        &provider,
+        gtk::STYLE_PROVIDER_PRIORITY_APPLICATION,
+    );
+}
+
+// ANCHOR: build_ui
 fn build_ui(app: &Application) {
     // ANCHOR: buttons
     // Create buttons
-    let button_1 = Button::with_label("Destructive");
-    let button_2 = Button::with_label("Suggested");
+    let button_1 = Button::with_label("Press me!");
+    let button_2 = Button::with_label("Press me!");
 
-    button_1.add_css_class("destructive-action");
-    button_2.add_css_class("suggested-action");
+    button_1.set_widget_name("button_1");
     // ANCHOR_END: buttons
 
     // Create `gtk_box` and add buttons
@@ -30,7 +46,7 @@ fn build_ui(app: &Application) {
         .margin_bottom(12)
         .margin_start(12)
         .margin_end(12)
-        .spacing(6)
+        .spacing(12)
         .build();
     gtk_box.append(&button_1);
     gtk_box.append(&button_2);
@@ -43,3 +59,4 @@ fn build_ui(app: &Application) {
         .build();
     window.show();
 }
+// ANCHOR_END: build_ui
