+++ book/listings/css/7/style.css
@@ -1,4 +1,4 @@
-button.button_1:hover {
+button#button-1:hover {
   color: magenta;
   background: yellow;
 }
