+++ book/listings/css/7/window/window.ui
@@ -9,6 +9,7 @@
         <property name="margin-bottom">12</property>
         <property name="margin-start">12</property>
         <property name="margin-end">12</property>
+        <property name="name">button_1</property>
         <style>
           <class name="button_1"/>
         </style>
