+++ book/listings/todo_app/4/todo_row/imp.rs
@@ -27,7 +27,6 @@ impl ObjectSubclass for TodoRow {
 
     fn class_init(klass: &mut Self::Class) {
         Self::bind_template(klass);
-        //klass.set_css_name("todo_row");
     }
 
     fn instance_init(obj: &glib::subclass::InitializingObject<Self>) {
